Solve now at least partially working, but slow
The solver is now working on at least simple sudoku puzzles,
but is very slow, and is either prohibitively slow at or
incapable of solving all but the most simple of puzzles.

Currently work is being done to profile the generated code
to find any bottlenecks in the implementation, and possible
changes are being tested.